refactor: link groups and customize collapsed views
